Account for newly enqueued jobs when returning session progress
Previously, if 10 jobs were queued when the session started, we'd run 10
jobs then stop since we thought the queue would be exhausted. If jobs
were enqueued during the session, they'd be left for the next session.

Now the progress will adapt to newly enqueued jobs and hopefully the
session wont end until the queue is truly empty, regardless of it's size
at the start of the session.